fix(core): Address third-round review on compression output budget clamp
Close the floor-regime hole on the usage-missing path: at a floored
budget of 1 the truncation guard now drops locally estimated output too,
since no complete summary can exist at a 1-token cap. Add regression
coverage for both the floor + missing-usage combination and the
provenance split (clamped budget + complete estimated summary). Correct
the safety-margin docstring to not overstate its guarantee against
proportional estimator error. Fold the issue-7960 regression suite into
the colocated main test file per repo convention.
